Moved SimpleJSON into 3rdparty
