+++ 3rdparty/CMakeLists.txt
@@ -1,1 +1,2 @@
 add_subdirectory(spdlog)
+add_subdirectory(SimpleJSON)
