+++ 3rdparty/SimpleJSON/JSON.cpp
@@ -22,7 +22,7 @@
  * THE SOFTWARE.
  */
 
-#include "JSON.h"
+#include <SimpleJSON/JSON.h>
 
 /**
  * Blocks off the public constructor
@@ -48,7 +48,7 @@ JSONValue *JSON::Parse(const char *data)
 {
 	size_t length = strlen(data) + 1;
 	wchar_t *w_data = (wchar_t*)malloc(length * sizeof(wchar_t));
-	
+
 	#if defined(WIN32) && !defined(__GNUC__)
 		size_t ret_value = 0;
 		if (mbstowcs_s(&ret_value, w_data, length, data, length) != 0)
@@ -67,7 +67,7 @@ JSONValue *JSON::Parse(const char *data)
 			return NULL;
 		}
 	#endif
-	
+
 	JSONValue *value = JSON::Parse(w_data);
 	free(w_data);
 	return value;
@@ -92,14 +92,14 @@ JSONValue *JSON::Parse(const wchar_t *data)
 	JSONValue *value = JSONValue::Parse(&data);
 	if (value == NULL)
 		return NULL;
-	
+
 	// Can be white space now and should be at the end of the string then...
 	if (SkipWhitespace(&data))
 	{
 		delete value;
 		return NULL;
 	}
-	
+
 	// We're now at the end of the string
 	return value;
 }
@@ -134,7 +134,7 @@ bool JSON::SkipWhitespace(const wchar_t **data)
 {
 	while (**data != 0 && (**data == L' ' || **data == L'\t' || **data == L'\r' || **data == L'\n'))
 		(*data)++;
-	
+
 	return **data != 0;
 }
 
@@ -152,18 +152,18 @@ bool JSON::SkipWhitespace(const wchar_t **data)
 bool JSON::ExtractString(const wchar_t **data, std::wstring &str)
 {
 	str = L"";
-	
+
 	while (**data != 0)
 	{
 		// Save the char so we can change it if need be
 		wchar_t next_char = **data;
-		
+
 		// Escaping something?
 		if (next_char == L'\\')
 		{
 			// Move over the escape char
 			(*data)++;
-			
+
 			// Deal with the escaped char
 			switch (**data)
 			{
@@ -180,7 +180,7 @@ bool JSON::ExtractString(const wchar_t **data, std::wstring &str)
 					// We need 5 chars (4 hex + the 'u') or its not valid
 					if (!simplejson_wcsnlen(*data, 5))
 						return false;
-					
+
 					// Deal with the chars
 					next_char = 0;
 					for (int i = 0; i < 4; i++)
@@ -188,9 +188,9 @@ bool JSON::ExtractString(const wchar_t **data, std::wstring &str)
 						// Do it first to move off the 'u' and leave us on the
 						// final hex digit as we move on by one later on
 						(*data)++;
-						
+
 						next_char <<= 4;
-						
+
 						// Parse the hex digit
 						if (**data >= '0' && **data <= '9')
 							next_char |= (**data - '0');
@@ -206,13 +206,13 @@ bool JSON::ExtractString(const wchar_t **data, std::wstring &str)
 					}
 					break;
 				}
-				
+
 				// By the spec, only the above cases are allowed
 				default:
 					return false;
 			}
 		}
-		
+
 		// End of the string?
 		else if (next_char == L'"')
 		{
@@ -220,21 +220,21 @@ bool JSON::ExtractString(const wchar_t **data, std::wstring &str)
 			str.reserve(); // Remove unused capacity
 			return true;
 		}
-		
+
 		// Disallowed char?
 		else if (next_char < L' ' && next_char != L'\t')
 		{
 			// SPEC Violation: Allow tabs due to real world cases
 			return false;
 		}
-		
+
 		// Add the next char
 		str += next_char;
-		
+
 		// Move on
 		(*data)++;
 	}
-	
+
 	// If we're here, the string ended incorrectly
 	return false;
 }
@@ -253,7 +253,7 @@ double JSON::ParseInt(const wchar_t **data)
 	double integer = 0;
 	while (**data != 0 && **data >= '0' && **data <= '9')
 		integer = integer * 10 + (*(*data)++ - '0');
-	
+
 	return integer;
 }
